Upgrade modeljson ints to int64 if their proto uses int64 (#122)
* introduce a nullable.Int64

* upgrade modeljson int to int64 if their proto uses int64

* remove transforms from int64 to int

* handle int64 in generator

* populate int64

+++ model/internal/modeljson/metricset.go
@@ -100,7 +100,7 @@ func (ms *MetricsetSample) MarshalFastJSON(w *fastjson.Writer) error {
 }
 
 type AggregatedDuration struct {
-	Count int
+	Count int64
 	Sum   time.Duration
 }
